fix(build): resolve all CI compilation errors from typed ops PRs
Fix 5 categories of build errors introduced by PRs 2-6:
1. Missing imports: dtype_ordinal moved to shared/base/ — update stale
   relative imports in activation.mojo, conv.mojo, matrix.mojo, reduction.mojo
2. Unqualified dtype -> Self.dtype in parametric struct methods for
   tensor.mojo, batchnorm.mojo, conv2d.mojo, linear.mojo
3. Argument mismatch in pooling typed wrappers: align kernel_size param
   with AnyTensor API (was kernel_h/kernel_w causing Int->String error)
4. Immutable capture in SIMD closures: extract data pointers into local
   vars before closure capture in activation_simd.mojo (5 functions)
5. Missing Tensor import in any_tensor.mojo (top-level instead of
   function-scoped to resolve return type annotation)
6. Float64->Float32 type mismatch in smooth_l1_loss_typed beta param

+++ shared/core/loss.mojo
@@ -1120,7 +1120,7 @@ fn smooth_l1_loss_typed[
 ](
     predictions: Tensor[dt],
     targets: Tensor[dt],
-    beta: Float64 = 1.0,
+    beta: Float32 = 1.0,
 ) raises -> Tensor[dt]:
     """Typed overload of smooth_l1_loss for Tensor[dtype].
